pyproject.toml
[tool.poetry]
name = "mtuci-telegram"
version = "0.1.0"
description = ""
authors = ["Daniel <matsubus@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
pyTelegramBotAPI = "^4.4.0"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

